\name{hdr}
\alias{hdr}
\docType{data}
\title{Deceased in Slovenia in 1998}
\description{
  Personal income tax payed (in SIT), age (in years) and gender for deceased in Slovenia in 1998.
}
\usage{data(hdr)}
\format{
\code{hdr} is a list with 9051 cases (rows) and 3 variables (columns) named \code{income}, \code{age} and \code{gender}.
}
\source{Statistical Office of the Republic of Slovenia.}
\references{
Javornik, J., and Korosec, V. Eds. (2003) \emph{Human development report Slovenia 2002/2003: Human development and health}.
Ljubljana: Institute of Macroecomic Analysis and Development.
}

\keyword{datasets}
